Create Godot Project, initial exploration/setup
The most notable addition is the "Configuration" node/scene, which is set to be auto loaded, and thus be a hub for things like what colors the numbers can be. Since I plan to expand on this project after the jam, (or rather make another project with the same core mechanics), this kind of set up will allow an options menu where colors can be configured by the user rather than just me, improving accessibility.